Move designer/client test helpers into designer/client
