+++ designer/client/jest.environment.js
@@ -1,5 +1,5 @@
 require('@testing-library/jest-dom')
-require('../test/testServer')
+require('./test/testServer')
 const { initI18n } = require('./src/i18n')
 
 initI18n()
