+++ designer/client/src/components/Autocomplete/Autocomplete.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { screen } from '@testing-library/dom'
 import { cleanup, render } from '@testing-library/react'
 import { Autocomplete } from './Autocomplete'
-import { RenderWithContext } from '../../../../test/helpers/renderers'
+import { RenderWithContext } from '../../../test/helpers/renderers'
 
 describe('AutocompleteField', () => {
   const { getByText } = screen
