+++ designer/client/src/components/CustomValidationMessage/CustomValidationMessage.test.tsx
@@ -3,7 +3,7 @@ import { screen } from '@testing-library/dom'
 import { act, cleanup, render } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import { CustomValidationMessage } from './CustomValidationMessage'
-import { RenderWithContext } from '../../../../test/helpers/renderers'
+import { RenderWithContext } from '../../../test/helpers/renderers'
 
 describe('CssClasses', () => {
   afterEach(cleanup)
