+++ designer/client/src/components/FieldEditors/number-field-edit.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { screen } from '@testing-library/dom'
 import { cleanup, render } from '@testing-library/react'
 import { NumberFieldEdit } from './number-field-edit'
-import { RenderWithContext } from '../../../../test/helpers/renderers'
+import { RenderWithContext } from '../../../test/helpers/renderers'
 
 describe('Number field edit', () => {
   afterEach(cleanup)
