+++ designer/client/src/components/FieldEditors/text-field-edit.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { screen } from '@testing-library/dom'
 import { render } from '@testing-library/react'
 import { TextFieldEdit } from './text-field-edit'
-import { RenderWithContext } from '../../../../test/helpers/renderers'
+import { RenderWithContext } from '../../../test/helpers/renderers'
 
 describe('Text field edit', () => {
   const { getByText } = screen
