+++ designer/client/src/load-form-configurations.test.tsx
@@ -1,6 +1,6 @@
 import { http } from 'msw'
 import * as formConfigurationsApi from './load-form-configurations'
-import { server, mockedFormHandlers } from '../../test/testServer'
+import { server, mockedFormHandlers } from '../test/testServer'
 
 describe('Load form configurations', () => {
   beforeAll(() => server.listen())
