+++ designer/client/src/outputs/output-edit.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { screen } from '@testing-library/dom'
 import { act, cleanup, render, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { RenderWithContextAndDataContext } from '../../../test/helpers/renderers'
+import { RenderWithContextAndDataContext } from '../../test/helpers/renderers'
 import { Data } from '@defra/forms-model'
 
 import OutputEdit from './output-edit'
