+++ designer/client/src/pages/LandingPage/NewConfig.test.tsx
@@ -4,7 +4,7 @@ import { act, cleanup, render, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import React from 'react'
 import { MemoryRouter } from 'react-router-dom'
-import { server, http, mockedFormHandlers } from '../../../../test/testServer'
+import { server, http, mockedFormHandlers } from '../../../test/testServer'
 import type { FormConfiguration } from '@defra/forms-model'
 
 describe('Newconfig', () => {
