+++ designer/client/src/section/section-edit.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { screen } from '@testing-library/dom'
 import { cleanup, render } from '@testing-library/react'
-import { RenderWithContext } from '../../../test/helpers/renderers'
+import { RenderWithContext } from '../../test/helpers/renderers'
 import SectionEdit from './section-edit'
 
 describe('Section edit fields', () => {
