+++ designer/client/test/helpers/renderers.tsx
@@ -3,8 +3,8 @@ import {
   ComponentContext,
   componentReducer,
   initComponentState
-} from '../../client/src/reducers/component/componentReducer'
-import { DataContext } from '../../client/src/context'
+} from '../../src/reducers/component/componentReducer'
+import { DataContext } from '../../src/context'
 
 export function RenderWithContext({ children, stateProps = {} }) {
   const [state, dispatch] = useReducer(
